v0.3: dashboard refactor — drop Inbox, fold Peers, split AgentView monolith
Six discrete UI changes per master plan §4 PR-8 + p4 dashboard refactor:

1. Inbox tab DELETED (approvals UX now via sidebar pip from PR-6 + future
   inline approval cards in PR-10 HermesChat).
2. Peers tab folded into Memory tab as "Peer memory" subsection — the
   live MCP search Peers used (R5 finding) is preserved, not deleted.
3. AgentView 974-LOC monolith split into ui/src/dash/agents/{agent-view,
   hermes-chat-tab,personas-tab,skills-tab,memory-tab,plugins-tab}.jsx.
4. HermesChatTab is now the default tab (placeholder; PR-10 fills in
   composer + transcript).
5. data.jsx purged of agent-related mock entries (HAL0_DATA.approvals).
6. Old test.skip-only agent-v3.spec.ts deleted; new minimal smoke spec
   agent-view-v3.spec.ts covers nav + default tab + Inbox/Peers removal +
   #peers legacy redirect.

Window-globals build shim preserved. Backend untouched.

Refs MASTER-PLAN.md §4 PR-8.
