+++ ui/src/dash/extras.jsx
@@ -1,12 +1,13 @@
-// hal0 dashboard — secondary views: Hardware, Logs, Backends, Agent
+// hal0 dashboard — secondary views: Backends, Logs.
 //
-// Phase B1: Hardware / Backends / Logs read from real hooks. AgentView
-// stays on HAL0_DATA mock (deferred; follow-up issue tracks Phase B2 +).
+// Phase B1: Backends + Logs read from real hooks. v0.3 PR-8 split the
+// AgentView monolith out of this file into ui/src/dash/agents/* — see
+// agent-view.jsx, hermes-chat-tab.jsx, personas-tab.jsx, skills-tab.jsx,
+// memory-tab.jsx, plugins-tab.jsx.
 
 import { useBackends } from '@/api/hooks/useBackends'
 import { useLogsHistorical, useLogsStream } from '@/api/hooks/useLogs'
 import { useLemondRollup } from '@/api/hooks/useLemonade'
-import { useMemoryGraphStatus, useUpdateMemoryGraph } from '@/api/hooks/useMemory'
 
 const { useState: useStateX } = React;
 
@@ -377,735 +378,4 @@ function highlightSearch(text, q) {
   );
 }
 
-// ════════════════════════════════════════════════════════════════════
-// AGENT (chat, skills, memory, personas)
-// ════════════════════════════════════════════════════════════════════
-function AgentView() {
-  const [tab, setTab] = useStateX("overview");
-  const [editPersona, setEditPersona] = useStateX(null);
-  const [resetOpen, setResetOpen] = useStateX(false);
-  const noAgent = window.__hal0Banners && window.__hal0Banners.get && window.__hal0Banners.get()["no-agent"];
-  const tabs = [
-    { id: "overview", label: "Overview" },
-    { id: "inbox",    label: "Inbox" },
-    { id: "skills",   label: "Skills" },
-    { id: "memory",   label: "Memory" },
-    { id: "personas", label: "Personas" },
-    { id: "peers",    label: "Peers" },
-    // v0.3 PR-7: plugin host mounts upstream Hermes plugin tabs
-    // (kanban today, future plugins free) under a single nav entry.
-    // PR-8 splits AgentView; this is the minimal edit until then.
-    { id: "plugins",  label: "Plugins" },
-  ];
-  return (
-    <div className="view">
-      <div className="vh">
-        <span className="vh-eye mono">Tools</span>
-        <h1>Agent</h1>
-        <span className="vh-spacer" />
-        <span className="hint mono">scaffolded · v0.2.1 · full surface in v0.3</span>
-      </div>
-
-      <div style={{display: "flex", gap: 0, borderBottom: "1px solid var(--line)", marginBottom: 18}}>
-        {tabs.map(t => (
-          <button
-            key={t.id}
-            onClick={() => setTab(t.id)}
-            style={{
-              padding: "10px 16px",
-              background: "transparent",
-              border: "none",
-              borderBottom: tab === t.id ? "2px solid var(--accent)" : "2px solid transparent",
-              color: tab === t.id ? "var(--accent)" : "var(--fg-3)",
-              fontFamily: "var(--jbm)",
-              fontSize: 12.5,
-              cursor: "pointer",
-              fontWeight: 500,
-            }}
-          >{t.label}</button>
-        ))}
-      </div>
-
-      {tab === "overview" && (noAgent ? <NoBundledAgentCard /> : <AgentOverview />)}
-      {tab === "inbox"    && (noAgent ? <EmptyInbox /> : <AgentInbox />)}
-      {tab === "skills"   && <AgentSkills />}
-      {tab === "memory"   && <AgentMemory onResetNs={() => setResetOpen(true)} />}
-      {tab === "personas" && <AgentPersonas onEdit={(p) => setEditPersona(p)} />}
-      {tab === "peers"    && <AgentPeers />}
-      {tab === "plugins"  && window.PluginTabHost && <window.PluginTabHost agentId="hermes" />}
-
-      <PersonaEditModal
-        open={!!editPersona}
-        persona={editPersona}
-        onClose={() => setEditPersona(null)}
-      />
-      <ConfirmDialog
-        open={resetOpen}
-        onCancel={() => setResetOpen(false)}
-        onConfirm={() => { setResetOpen(false); window.__hal0Toast && window.__hal0Toast("Cognee namespace 'shared' reset — 2,847 records deleted", "warn"); }}
-        title="Reset memory namespace 'shared'?"
-        message={<span>This permanently deletes <span className="mono" style={{color: "var(--fg)"}}>2,847</span> Cognee records across SQLite + LanceDB + Kuzu. Cannot be undone.</span>}
-        confirmLabel="Reset namespace"
-        destructive
-        typeToConfirm="shared"
-      />
-    </div>
-  );
-}
-
-// Empty inbox state (no agent installed)
-function EmptyInbox() {
-  return (
-    <div className="card" style={{padding: 40, textAlign: "center", borderStyle: "dashed"}}>
-      <div className="mono" style={{fontSize: 14, color: "var(--fg-3)", marginBottom: 6}}>No pending approvals.</div>
-      <div className="mono" style={{fontSize: 11, color: "var(--fg-5)"}}>Gated tool calls from agents will appear here once an agent is installed.</div>
-    </div>
-  );
-}
-
-function AgentOverview() {
-  return (
-    <div style={{display: "grid", gridTemplateColumns: "1fr 320px", gap: 16}}>
-      <div>
-        <div className="sec">
-          <h2>Bundled agent</h2>
-          <div className="rule" />
-        </div>
-        <div className="card" style={{padding: 22, marginBottom: 18}}>
-          <div style={{display: "flex", alignItems: "center", gap: 14, marginBottom: 14}}>
-            <div style={{width: 44, height: 44, borderRadius: 8, background: "var(--accent-soft)", border: "1px solid var(--accent-line)", display: "inline-flex", alignItems: "center", justifyContent: "center"}}>
-              {Icons.agent}
-            </div>
-            <div>
-              <div className="mono" style={{fontSize: 16, fontWeight: 500, letterSpacing: "-0.02em"}}>Hermes-Agent</div>
-              <div className="mono" style={{fontSize: 11, color: "var(--fg-3)", marginTop: 2}}>service · hal0-agent-hermes.service · running</div>
-            </div>
-            <span style={{marginLeft: "auto"}} className="chip ok">running · 14d</span>
-          </div>
-          <p style={{fontSize: 13, color: "var(--fg-2)", margin: "0 0 16px", maxWidth: 640, lineHeight: 1.55}}>
-            Hermes is the resident agent. It exposes a chat surface, executes skills via MCP, writes to Cognee memory, and respects the approval policy from Settings. It runs as a systemd service for v0.2.1; CLI shape (pi-coder) lands in v0.3.
-          </p>
-          <div style={{display: "grid", gridTemplateColumns: "repeat(4, 1fr)", gap: 0, border: "1px solid var(--line)", borderRadius: "var(--rad)", overflow: "hidden"}}>
-            {[
-              { l: "approvals", v: "3", sub: "pending" },
-              { l: "skills",    v: "12", sub: "wired" },
-              { l: "memory",    v: "847", sub: "writes" },
-              { l: "persona",   v: "hermes", sub: "default" },
-            ].map((s, i) => (
-              <div key={i} style={{padding: 14, borderRight: i < 3 ? "1px solid var(--line)" : "none"}}>
-                <div className="mono" style={{fontSize: 10, color: "var(--fg-4)", textTransform: "uppercase", letterSpacing: "0.08em"}}>{s.l}</div>
-                <div className="mono num" style={{fontSize: 20, color: "var(--fg)", letterSpacing: "-0.02em", marginTop: 4}}>{s.v}</div>
-                <div className="mono" style={{fontSize: 10, color: "var(--fg-4)"}}>{s.sub}</div>
-              </div>
-            ))}
-          </div>
-          <div style={{marginTop: 14, display: "flex", gap: 8}}>
-            <button className="btn">{Icons.send} Open chat with Hermes</button>
-            <button className="btn ghost">{Icons.logs} View activity</button>
-            <button className="btn ghost">{Icons.restart} Restart service</button>
-          </div>
-        </div>
-
-        <div className="sec">
-          <h2>Alternative: pi-coder<span className="ct mono">CLI · v0.3</span></h2>
-          <div className="rule" />
-        </div>
-        <div className="card" style={{padding: 18, opacity: 0.7}}>
-          <div className="mono" style={{fontSize: 13, marginBottom: 6}}>@earendil-works/pi-coding-agent</div>
-          <div className="mono" style={{fontSize: 11, color: "var(--fg-3)", marginBottom: 12}}>CLI shape · 4 tools · invoked per-task · not installed</div>
-          <button className="btn ghost sm" disabled>Install pi-coder (v0.3)</button>
-        </div>
-      </div>
-
-      <div>
-        <div className="side-card">
-          <div className="side-card-h">
-            <span>Recent activity</span>
-          </div>
-          <div className="side-card-b">
-            {[
-              { ts: "14:02:09", text: "request model_pull user.Phi-4-Mini", st: "pending" },
-              { ts: "14:01:42", text: "wrote /scratch/notes/draft.md", st: "approved" },
-              { ts: "14:00:18", text: "shell rg \"TODO\" src/", st: "pending" },
-              { ts: "13:58:11", text: "wrote cognee record (847)", st: "auto" },
-              { ts: "13:54:33", text: "read src/hal0/launchers/", st: "auto" },
-              { ts: "13:49:02", text: "denied: shell rm -rf /tmp/cache", st: "denied" },
-            ].map((a, i) => (
-              <div key={i} style={{padding: "8px 0", borderBottom: "1px solid var(--line-soft)", display: "grid", gridTemplateColumns: "64px 1fr 70px", gap: 8, fontFamily: "var(--jbm)", fontSize: 11}}>
-                <span style={{color: "var(--fg-5)"}}>{a.ts}</span>
-                <span style={{color: "var(--fg-2)", overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap"}}>{a.text}</span>
-                <span style={{textAlign: "right"}}>
-                  {a.st === "pending"  && <span className="chip warn">pending</span>}
-                  {a.st === "approved" && <span className="chip ok">ok</span>}
-                  {a.st === "auto"     && <span className="chip">auto</span>}
-                  {a.st === "denied"   && <span className="chip err">denied</span>}
-                </span>
-              </div>
-            ))}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function AgentInbox() {
-  return (
-    <div>
-      <div className="card" style={{padding: 16, marginBottom: 14, display: "flex", alignItems: "center", gap: 12, borderColor: "var(--warn-line)", background: "var(--warn-soft)"}}>
-        <span style={{color: "var(--warn)"}}>{Icons.warn}</span>
-        <span className="mono" style={{fontSize: 12, color: "var(--warn)"}}>
-          <b>3 pending approvals.</b> Hermes is paused on these calls until you resolve them.
-        </span>
-        <span className="mono" style={{marginLeft: "auto", fontSize: 11, color: "var(--fg-4)"}}>policy: registry-write · always</span>
-      </div>
-      {HAL0_DATA.approvals.map((a, i) => (
-        <div key={i} className="card" style={{padding: 0, marginBottom: 10, overflow: "hidden"}}>
-          <div style={{padding: "12px 16px", borderBottom: "1px solid var(--line-soft)", display: "flex", alignItems: "center", gap: 12, fontFamily: "var(--jbm)", fontSize: 12}}>
-            <span style={{color: "var(--fg-4)"}}>{a.ts}</span>
-            <span style={{color: "var(--accent)", fontWeight: 500}}>{a.agent}</span>
-            <span style={{color: "var(--fg-4)"}}>requests</span>
-            <span style={{color: "var(--fg)", fontWeight: 500}}>{a.tool}</span>
-            <span style={{marginLeft: "auto"}}>
-              <span className="chip warn">awaiting approval · 14s</span>
-            </span>
-          </div>
-          <div style={{padding: "14px 16px", fontFamily: "var(--jbm)", fontSize: 12, display: "grid", gap: 6}}>
-            <div style={{display: "grid", gridTemplateColumns: "120px 1fr", gap: 12}}>
-              <span style={{color: "var(--fg-4)"}}>argument</span>
-              <span style={{color: "var(--fg-2)"}}>{a.arg}</span>
-            </div>
-            <div style={{display: "grid", gridTemplateColumns: "120px 1fr", gap: 12}}>
-              <span style={{color: "var(--fg-4)"}}>capability</span>
-              <span style={{color: "var(--fg-2)"}}>{a.tool.startsWith("model_") ? "registry-write" : a.tool.startsWith("fs_") ? "fs-write" : "shell-exec"}</span>
-            </div>
-            <div style={{display: "grid", gridTemplateColumns: "120px 1fr", gap: 12}}>
-              <span style={{color: "var(--fg-4)"}}>reason</span>
-              <span style={{color: "var(--fg-2)"}}>{a.tool === "model_pull" ? "user asked Hermes to set up Phi-4-Mini for offline routing" : a.tool === "fs_write" ? "draft notes from current chat session" : "code search across src/"}</span>
-            </div>
-          </div>
-          <div style={{padding: "10px 16px", borderTop: "1px solid var(--line-soft)", background: "var(--bg)", display: "flex", gap: 6, justifyContent: "flex-end"}}>
-            <button className="btn danger sm">Deny</button>
-            <button className="btn ghost sm">Deny + remember</button>
-            <button className="btn ghost sm">Approve once</button>
-            <button className="btn sm">Approve + remember</button>
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function AgentSkills() {
-  const skills = [
-    { name: "read_file",       cap: "fs-read",      policy: "remember", calls: 247, src: "builtin" },
-    { name: "write_file",      cap: "fs-write",     policy: "always",   calls: 38,  src: "builtin" },
-    { name: "edit_file",       cap: "fs-write",     policy: "always",   calls: 14,  src: "builtin" },
-    { name: "list_dir",        cap: "fs-read",      policy: "remember", calls: 41,  src: "builtin" },
-    { name: "shell_exec",      cap: "shell-exec",   policy: "always",   calls: 9,   src: "builtin" },
-    { name: "model_pull",      cap: "registry-write", policy: "always", calls: 3,   src: "hal0-router" },
-    { name: "restart_slot",    cap: "slot-control", policy: "always",   calls: 1,   src: "hal0-router" },
-    { name: "generate_image",  cap: "tool-call",    policy: "auto",     calls: 18,  src: "omnirouter" },
-    { name: "transcribe_audio",cap: "tool-call",    policy: "auto",     calls: 7,   src: "omnirouter" },
-    { name: "text_to_speech",  cap: "tool-call",    policy: "auto",     calls: 22,  src: "omnirouter" },
-    { name: "embed_text",      cap: "tool-call",    policy: "auto",     calls: 184, src: "omnirouter" },
-    { name: "rerank_documents",cap: "tool-call",    policy: "auto",     calls: 41,  src: "omnirouter" },
-  ];
-  return (
-    <div>
-      <div className="card" style={{overflow: "hidden"}}>
-        <div style={{padding: "10px 18px", background: "var(--bg)", borderBottom: "1px solid var(--line)", display: "grid", gridTemplateColumns: "200px 160px 1fr 120px 80px auto", gap: 16, fontFamily: "var(--jbm)", fontSize: 10, color: "var(--fg-4)", textTransform: "uppercase", letterSpacing: "0.08em"}}>
-          <span>skill</span>
-          <span>capability</span>
-          <span>source</span>
-          <span>policy</span>
-          <span style={{textAlign: "right"}}>calls</span>
-          <span></span>
-        </div>
-        {skills.map(s => (
-          <div key={s.name} style={{padding: "11px 18px", borderBottom: "1px solid var(--line-soft)", display: "grid", gridTemplateColumns: "200px 160px 1fr 120px 80px auto", gap: 16, alignItems: "center", fontFamily: "var(--jbm)", fontSize: 12}}>
-            <span style={{color: "var(--fg)", fontWeight: 500}}>{s.name}</span>
-            <span style={{color: "var(--fg-3)"}}>{s.cap}</span>
-            <span style={{color: "var(--fg-4)"}}>{s.src}</span>
-            <span>
-              {s.policy === "always"   && <span className="chip warn">always</span>}
-              {s.policy === "remember" && <span className="chip ok">remember</span>}
-              {s.policy === "auto"     && <span className="chip">auto</span>}
-              {s.policy === "deny"     && <span className="chip err">deny</span>}
-            </span>
-            <span style={{textAlign: "right", color: "var(--fg-2)"}} className="num">{s.calls}</span>
-            <button className="btn ghost sm">{Icons.edit}</button>
-          </div>
-        ))}
-      </div>
-      <div style={{marginTop: 14, fontFamily: "var(--jbm)", fontSize: 11, color: "var(--fg-4)"}}>
-        12 skills wired · 8 require approval · 4 auto via OmniRouter · skill source includes builtin, hal0-router, omnirouter, and any user-added MCP servers (none configured).
-      </div>
-    </div>
-  );
-}
-
-function AgentMemory({ onResetNs }) {
-  return (
-    <div style={{display: "grid", gridTemplateColumns: "1fr 320px", gap: 16}}>
-      <div>
-        <GraphExtractionPanel />
-        <div className="card" style={{padding: 18, marginBottom: 14}}>
-          <div style={{display: "flex", alignItems: "center", gap: 12, marginBottom: 14}}>
-            <span className="mono" style={{fontSize: 10, color: "var(--accent)", textTransform: "uppercase", letterSpacing: "0.1em"}}>Cognee · shared</span>
-            <span className="mono num" style={{fontSize: 24, color: "var(--fg)", letterSpacing: "-0.02em"}}>2,847</span>
-            <span className="mono" style={{fontSize: 12, color: "var(--fg-3)"}}>records · 184 MB</span>
-            <span style={{marginLeft: "auto"}} className="chip ok">healthy</span>
-          </div>
-          <div style={{display: "grid", gridTemplateColumns: "repeat(3, 1fr)", gap: 0, border: "1px solid var(--line)", borderRadius: "var(--rad)", overflow: "hidden"}}>
-            {[
-              { l: "SQLite", v: "847", sub: "indexed text" },
-              { l: "LanceDB", v: "2,140", sub: "vectors · 768d" },
-              { l: "Kuzu",   v: "412", sub: "graph edges" },
-            ].map((s, i) => (
-              <div key={i} style={{padding: 14, borderRight: i < 2 ? "1px solid var(--line)" : "none"}}>
-                <div className="mono" style={{fontSize: 10, color: "var(--fg-4)", textTransform: "uppercase", letterSpacing: "0.08em"}}>{s.l}</div>
-                <div className="mono num" style={{fontSize: 22, color: "var(--fg)", marginTop: 4}}>{s.v}</div>
-                <div className="mono" style={{fontSize: 10, color: "var(--fg-4)"}}>{s.sub}</div>
-              </div>
-            ))}
-          </div>
-        </div>
-
-        <div className="sec"><h2>Recent records</h2><div className="rule" /></div>
-        <div className="card" style={{overflow: "hidden"}}>
-          {[
-            { ts: "14:02:11", source: "hermes", kind: "fact",       body: "user prefers frozen dataclasses for SlotState types" },
-            { ts: "14:00:42", source: "hermes", kind: "convo",       body: "session ftr-104 — refactor of slot_manager.py" },
-            { ts: "13:58:18", source: "hermes", kind: "code-ref",    body: "slot.py:42 — SlotState dataclass with slots=True" },
-            { ts: "13:54:01", source: "hermes", kind: "skill-trace", body: "read_file → src/hal0/launchers/slot_manager.py (3 calls)" },
-            { ts: "13:50:33", source: "user",   kind: "preference",  body: "models page: sort installed first" },
-          ].map((r, i) => (
-            <div key={i} style={{padding: "12px 18px", borderBottom: "1px solid var(--line-soft)", fontFamily: "var(--jbm)", fontSize: 12}}>
-              <div style={{display: "flex", gap: 10, marginBottom: 4}}>
-                <span style={{color: "var(--fg-5)"}}>{r.ts}</span>
-                <span style={{color: "var(--accent)"}}>{r.source}</span>
-                <span className="chip">{r.kind}</span>
-              </div>
-              <div style={{color: "var(--fg-2)", paddingLeft: 0}}>{r.body}</div>
-            </div>
-          ))}
-        </div>
-      </div>
-
-      <div>
-        <div className="side-card">
-          <div className="side-card-h"><span>Namespaces</span></div>
-          <div className="side-card-b">
-            {[
-              { name: "shared", desc: "default · all agents", recs: 2847, active: true },
-              { name: "scratch", desc: "ephemeral · auto-prune", recs: 84,  active: false },
-              { name: "code",    desc: "code refs only",       recs: 412, active: false },
-            ].map(n => (
-              <div key={n.name} style={{padding: "10px 0", borderBottom: "1px solid var(--line-soft)", display: "flex", alignItems: "center", gap: 10, fontFamily: "var(--jbm)", fontSize: 12}}>
-                <span className={"dot " + (n.active ? "ready" : "idle")} />
-                <div>
-                  <div style={{color: "var(--fg)", fontWeight: 500}}>{n.name}</div>
-                  <div style={{color: "var(--fg-4)", fontSize: 10}}>{n.desc}</div>
-                </div>
-                <span style={{marginLeft: "auto", color: "var(--fg-3)"}} className="num">{n.recs}</span>
-              </div>
-            ))}
-            <button className="btn ghost sm" style={{marginTop: 10, width: "100%", justifyContent: "center"}} onClick={() => window.__hal0Toast && window.__hal0Toast("New namespace modal — stubbed", "info")}>{Icons.plus} New namespace</button>
-            <button className="btn danger sm" style={{marginTop: 6, width: "100%", justifyContent: "center"}} onClick={onResetNs}>{Icons.warn} Reset 'shared'</button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// ── ADR-0014 Graph extraction panel ──────────────────────────────────
-// One toggle, one route picker, one disclosure, one error counter
-// (per ADR-0014 §3 + §4 + §6). Lives inside the Memory tab so the
-// rest of the memory surface (namespaces, recent records) renders
-// alongside without a route change.
-//
-// Disclosure copy MUST match ADR-0014 §3 + §4 verbatim — copy changes
-// require ADR amendment.
-function GraphExtractionPanel() {
-  const status = useMemoryGraphStatus();
-  const update = useUpdateMemoryGraph();
-  const data = status.data;
-  const enabled = !!(data && data.enabled);
-  const route = (data && data.route) || "upstream";
-
-  const [showRoutePicker, setShowRoutePicker] = useStateX(false);
-  const [draftRoute, setDraftRoute] = useStateX(route);
-  const [draftProvider, setDraftProvider] = useStateX(
-    (data && data.upstream && data.upstream.provider) || "openrouter",
-  );
-  const [draftModel, setDraftModel] = useStateX(
-    (data && data.upstream && data.upstream.model) || "anthropic/claude-3.5-sonnet",
-  );
-
-  const openPanel = () => {
-    setDraftRoute(route);
-    setDraftProvider((data && data.upstream && data.upstream.provider) || "openrouter");
-    setDraftModel((data && data.upstream && data.upstream.model) || "anthropic/claude-3.5-sonnet");
-    setShowRoutePicker(true);
-  };
-
-  const submit = () => {
-    const payload = { enabled: true, route: draftRoute };
-    if (draftRoute === "upstream") {
-      payload.upstream = { provider: draftProvider, model: draftModel };
-    }
-    update.mutate(payload, {
-      onSuccess: () => {
-        setShowRoutePicker(false);
-        window.__hal0Toast && window.__hal0Toast("Graph extraction enabled", "ok");
-      },
-      onError: (err) => {
-        window.__hal0Toast && window.__hal0Toast(`Enable failed: ${err.message}`, "err");
-      },
-    });
-  };
-
-  const disable = () => {
-    update.mutate({ enabled: false }, {
-      onSuccess: () => {
-        window.__hal0Toast && window.__hal0Toast("Graph extraction disabled", "warn");
-      },
-    });
-  };
-
-  if (status.isLoading) {
-    return (
-      <div className="card" style={{padding: 18, marginBottom: 14}}>
-        <div className="mono" style={{fontSize: 12, color: "var(--fg-4)"}}>Loading graph extraction status…</div>
-      </div>
-    );
-  }
-  if (status.isError) {
-    return (
-      <div className="card" style={{padding: 18, marginBottom: 14, borderColor: "var(--err-line, var(--line))"}}>
-        <div className="mono" style={{fontSize: 12, color: "var(--err, #c66)"}}>Memory engine unavailable: {String(status.error?.message || "unknown")}</div>
-      </div>
-    );
-  }
-
-  const errors = (data && data.errors) || 0;
-  const builds = (data && data.builds_ok) || 0;
-  const inFlight = (data && data.in_flight) || 0;
-  const lastBuilt = data && data.last_built_at;
-
-  return (
-    <div className="card" style={{padding: 18, marginBottom: 14}}>
-      <div style={{display: "flex", alignItems: "center", gap: 12, marginBottom: 10}}>
-        <span className="mono" style={{fontSize: 10, color: "var(--accent)", textTransform: "uppercase", letterSpacing: "0.1em"}}>Graph extraction · ADR-0014</span>
-        <span style={{marginLeft: "auto"}}>
-          {enabled ? <span className="chip ok">ON · {route}</span> : <span className="chip">OFF</span>}
-        </span>
-      </div>
-
-      {!enabled && !showRoutePicker && (
-        <>
-          <p style={{fontSize: 13, color: "var(--fg-2)", margin: "0 0 14px", lineHeight: 1.55}}>
-            Graph extraction is off. Memory still stores + searches vectors; the entity/relation graph powering <span className="mono" style={{color: "var(--fg)"}}>memory_search(mode="graph")</span> isn't built.
-          </p>
-          <button className="btn primary sm" onClick={openPanel}>
-            Enable graph extraction
-          </button>
-        </>
-      )}
-
-      {enabled && !showRoutePicker && (
-        <>
-          <div style={{display: "grid", gridTemplateColumns: "repeat(3, 1fr)", gap: 0, border: "1px solid var(--line)", borderRadius: "var(--rad)", overflow: "hidden", marginBottom: 12}}>
-            {[
-              { l: "Builds OK", v: String(builds), sub: "lifetime" },
-              { l: "Errors", v: String(errors), sub: errors ? "see logs" : "—" },
-              { l: "In flight", v: String(inFlight), sub: "pending" },
-            ].map((s, i) => (
-              <div key={i} style={{padding: 14, borderRight: i < 2 ? "1px solid var(--line)" : "none"}}>
-                <div className="mono" style={{fontSize: 10, color: "var(--fg-4)", textTransform: "uppercase", letterSpacing: "0.08em"}}>{s.l}</div>
-                <div className="mono num" style={{fontSize: 22, color: errors > 0 && i === 1 ? "var(--err, #c66)" : "var(--fg)", marginTop: 4}}>{s.v}</div>
-                <div className="mono" style={{fontSize: 10, color: "var(--fg-4)"}}>{s.sub}</div>
-              </div>
-            ))}
-          </div>
-          {lastBuilt && (
-            <div className="mono" style={{fontSize: 11, color: "var(--fg-4)", marginBottom: 10}}>
-              Last build: {lastBuilt}
-            </div>
-          )}
-          {data && data.last_error && (
-            <div className="mono" style={{fontSize: 11, color: "var(--err, #c66)", marginBottom: 10}}>
-              Last error: {data.last_error}
-            </div>
-          )}
-          <div style={{display: "flex", gap: 8}}>
-            <button className="btn ghost sm" onClick={openPanel}>{Icons.edit} Change route</button>
-            <button className="btn danger sm" onClick={disable}>Disable</button>
-          </div>
-        </>
-      )}
-
-      {showRoutePicker && (
-        <div style={{display: "flex", flexDirection: "column", gap: 14}}>
-          <div>
-            <div className="mono" style={{fontSize: 10, color: "var(--fg-4)", textTransform: "uppercase", letterSpacing: "0.08em", marginBottom: 6}}>Route</div>
-            {[
-              { id: "upstream", label: "upstream", desc: "Send memory text to a hosted provider per build. Best quality." },
-              { id: "primary",  label: "primary",  desc: "Use the live primary slot. Stays on-box. Quality depends on the model." },
-              { id: "agent",    label: "agent",    desc: "Use the agent slot (NPU). Stays on-box. Latency low; context constrained." },
-            ].map(o => (
-              <label key={o.id} style={{display: "flex", alignItems: "flex-start", gap: 8, padding: "6px 0", cursor: "pointer"}}>
-                <input
-                  type="radio"
-                  name="graph-route"
-                  value={o.id}
-                  checked={draftRoute === o.id}
-                  onChange={() => setDraftRoute(o.id)}
-                />
-                <span style={{flex: 1}}>
-                  <span className="mono" style={{fontSize: 12, color: "var(--fg)", fontWeight: 500}}>{o.label}</span>
-                  <span style={{display: "block", fontSize: 11.5, color: "var(--fg-3)", lineHeight: 1.4}}>{o.desc}</span>
-                </span>
-              </label>
-            ))}
-          </div>
-
-          {draftRoute === "upstream" && (
-            <div style={{display: "grid", gridTemplateColumns: "1fr 1fr", gap: 8}}>
-              <label>
-                <div className="mono" style={{fontSize: 10, color: "var(--fg-4)", marginBottom: 4}}>Provider</div>
-                <select
-                  value={draftProvider}
-                  onChange={e => setDraftProvider(e.target.value)}
-                  style={{width: "100%", padding: "6px 8px", background: "var(--bg-2)", border: "1px solid var(--line)", color: "var(--fg)", fontFamily: "var(--jbm)", fontSize: 12}}
-                >
-                  <option value="openrouter">openrouter</option>
-                  <option value="anthropic">anthropic</option>
-                  <option value="openai">openai</option>
-                  <option value="custom">custom</option>
-                </select>
-              </label>
-              <label>
-                <div className="mono" style={{fontSize: 10, color: "var(--fg-4)", marginBottom: 4}}>Model</div>
-                <input
-                  type="text"
-                  value={draftModel}
-                  onChange={e => setDraftModel(e.target.value)}
-                  placeholder="anthropic/claude-3.5-sonnet"
-                  style={{width: "100%", padding: "6px 8px", background: "var(--bg-2)", border: "1px solid var(--line)", color: "var(--fg)", fontFamily: "var(--jbm)", fontSize: 12}}
-                />
-              </label>
-            </div>
-          )}
-
-          {/* ADR-0014 §3 verbatim privacy disclosure copy. */}
-          <div style={{padding: 12, border: "1px solid var(--line)", borderRadius: "var(--rad)", background: "var(--bg-2)"}}>
-            <div className="mono" style={{fontSize: 10, color: "var(--fg-4)", textTransform: "uppercase", letterSpacing: "0.08em", marginBottom: 4}}>Privacy</div>
-            <p style={{fontSize: 12, color: "var(--fg-2)", margin: 0, lineHeight: 1.5}}>
-              Graph extraction sends ingested memory text to&nbsp;
-              <span className="mono" style={{color: "var(--fg)"}}>{draftRoute === "upstream" ? `${draftProvider} (${draftModel})` : draftRoute === "primary" ? "your primary slot" : "your agent slot"}</span>
-              {draftRoute === "upstream" ? " for entity + relation extraction. Your raw memory store stays local. Switch to a local slot to keep everything on-box (quality may vary on small models)." : ". This stays on-box. Quality may vary on small models."}
-            </p>
-          </div>
-
-          {/* ADR-0014 §4 verbatim quality caveat copy. */}
-          <div style={{padding: 12, border: "1px dashed var(--line)", borderRadius: "var(--rad)"}}>
-            <p style={{fontSize: 12, color: "var(--fg-3)", margin: 0, lineHeight: 1.5}}>
-              Graph quality varies by model. We don't currently measure it for you — your results may vary.
-            </p>
-          </div>
-
-          <div style={{display: "flex", gap: 8, justifyContent: "flex-end"}}>
-            <button className="btn ghost sm" onClick={() => setShowRoutePicker(false)}>Cancel</button>
-            <button className="btn primary sm" onClick={submit} disabled={update.isPending}>
-              {update.isPending ? "Saving…" : enabled ? "Save route" : "Enable graph extraction"}
-            </button>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function AgentPersonas({ onEdit }) {
-  const personas = [
-    { name: "hermes",        slot: "primary", model: "qwen3.6-27b-mtp",  tone: "operator", desc: "Default — terse, technical, runs skills aggressively. Wired to the dashboard chat surface.", active: true },
-    { name: "hermes-coder",  slot: "coder",   model: "qwen3-coder-30b", tone: "code-focused", desc: "Swaps in when the persona dropdown picks coder. Optimised for refactors and review." },
-    { name: "hermes-npu",    slot: "agent",   model: "gemma3:1b",       tone: "low-latency", desc: "NPU coresident · for short follow-ups while keeping voice+embed warm." },
-    { name: "+ custom",      slot: null,      model: "",                tone: "",            desc: "Add a persona — pick a chat slot, set a system prompt, and pick a skill set.", isAdd: true },
-  ];
-  return (
-    <div style={{display: "grid", gridTemplateColumns: "repeat(2, 1fr)", gap: 14}}>
-      {personas.map((p, i) => (
-        <div key={i} className="card" style={{padding: 18, position: "relative", borderColor: p.active ? "var(--accent-line)" : "var(--line)", borderStyle: p.isAdd ? "dashed" : "solid"}}>
-          {p.active && <div style={{position: "absolute", top: 0, left: 0, right: 0, height: 2, background: "var(--accent)"}} />}
-          <div style={{display: "flex", alignItems: "center", gap: 10, marginBottom: 10}}>
-            <div style={{width: 36, height: 36, borderRadius: 6, background: p.isAdd ? "var(--bg-2)" : "var(--accent-soft)", border: "1px solid " + (p.isAdd ? "var(--line)" : "var(--accent-line)"), display: "inline-flex", alignItems: "center", justifyContent: "center", color: p.isAdd ? "var(--fg-4)" : "var(--accent)"}}>
-              {p.isAdd ? Icons.plus : Icons.agent}
-            </div>
-            <div>
-              <div className="mono" style={{fontSize: 14, fontWeight: 500, letterSpacing: "-0.01em"}}>{p.name}</div>
-              {p.slot && <div className="mono" style={{fontSize: 11, color: "var(--fg-3)", marginTop: 2}}>routes to slot <b style={{color: "var(--accent)"}}>{p.slot}</b> · {p.model}</div>}
-            </div>
-            {p.active && <span style={{marginLeft: "auto"}} className="chip amber">active</span>}
-          </div>
-          <p style={{fontSize: 12.5, color: "var(--fg-2)", margin: "0 0 12px", lineHeight: 1.55}}>{p.desc}</p>
-          {!p.isAdd && (
-            <div style={{display: "flex", gap: 6, alignItems: "center"}}>
-              <span className="chip">{p.tone}</span>
-              <span style={{marginLeft: "auto", display: "flex", gap: 6}}>
-                <button className="btn ghost sm" onClick={() => onEdit && onEdit(p)}>{Icons.edit} Edit</button>
-                {!p.active && <button className="btn ghost sm" onClick={() => window.__hal0Toast && window.__hal0Toast(`Persona ${p.name} activated`, "ok")}>Activate</button>}
-              </span>
-            </div>
-          )}
-          {p.isAdd && (
-            <div style={{display: "flex", justifyContent: "flex-end"}}>
-              <button className="btn ghost sm" onClick={() => onEdit && onEdit(p)}>{Icons.plus} Create persona</button>
-            </div>
-          )}
-        </div>
-      ))}
-    </div>
-  );
-}
-
-// ── AgentPeers (#247) ───────────────────────────────────────────────────────
-//
-// Reads identity cards from the `agents` Cognee dataset (ADR-0011) via
-// the hal0-memory MCP. Renders one row per card with a TCP-ping
-// reachability dot (no persistent stored field — pinged on render).
-//
-// Cards are immutable per ADR-0011 §2; this panel is read-only.
-
-function AgentPeers() {
-  const [cards, setCards] = useStateX([]);
-  const [loading, setLoading] = useStateX(true);
-  const [err, setErr] = useStateX(null);
-
-  React.useEffect(() => {
-    let cancelled = false;
-    (async () => {
-      try {
-        // #302: REST shim at /api/memory/search instead of /mcp/memory.
-        // The streamable-HTTP MCP transport at /mcp/memory/mcp requires
-        // the initialize handshake — not doable from a fetch() oneshot.
-        const resp = await fetch("/api/memory/search", {
-          method: "POST",
-          headers: { "Content-Type": "application/json", "X-hal0-Agent": "hal0-dashboard" },
-          body: JSON.stringify({
-            query: "agent identity",
-            tags: ["agent-identity"],
-            dataset: "agents",
-            limit: 50,
-          }),
-        });
-        const data = await resp.json();
-        if (cancelled) return;
-        const items = (data && data.items) || [];
-        setCards(items);
-        setLoading(false);
-      } catch (e) {
-        if (!cancelled) {
-          setErr(String(e));
-          setLoading(false);
-        }
-      }
-    })();
-    return () => { cancelled = true; };
-  }, []);
-
-  if (loading) {
-    return <div className="card" style={{padding: 20, color: "var(--fg-3)"}}>Loading peers…</div>;
-  }
-  if (err) {
-    return <div className="card" style={{padding: 20, color: "var(--err)"}}>memory MCP unreachable: {err}</div>;
-  }
-  if (!cards.length) {
-    return (
-      <div className="card" style={{padding: 40, textAlign: "center", borderStyle: "dashed"}}>
-        <div className="mono" style={{fontSize: 14, color: "var(--fg-3)", marginBottom: 6}}>No agent identity cards published yet.</div>
-        <div className="mono" style={{fontSize: 11, color: "var(--fg-5)"}}>Cards appear here when a bundled agent finishes <code>hal0 agent bootstrap</code>.</div>
-      </div>
-    );
-  }
-  return (
-    <div style={{display: "grid", gridTemplateColumns: "repeat(auto-fill, minmax(300px, 1fr))", gap: 12}}>
-      {cards.map((c, i) => <PeerCard key={i} card={c} />)}
-    </div>
-  );
-}
-
-function PeerCard({ card }) {
-  const md = (card && card.metadata) || {};
-  const endpoint = md.endpoint || {};
-  const hs = md.hal0_state || {};
-  const roles = md.roles || [];
-  const [reach, setReach] = useStateX("checking");
-  const [expanded, setExpanded] = useStateX(false);
-
-  React.useEffect(() => {
-    let cancelled = false;
-    const url = endpoint.url;
-    if (!url) { setReach("none"); return; }
-    const ctrl = new AbortController();
-    const tid = setTimeout(() => ctrl.abort(), 5000);
-    (async () => {
-      try {
-        await fetch(url, { method: "HEAD", signal: ctrl.signal, mode: "no-cors" });
-        if (!cancelled) setReach("ok");
-      } catch (e) {
-        if (cancelled) return;
-        setReach(e.name === "AbortError" ? "timeout" : "error");
-      } finally {
-        clearTimeout(tid);
-      }
-    })();
-    return () => { cancelled = true; ctrl.abort(); };
-  }, [endpoint.url]);
-
-  const dotColor = reach === "ok" ? "var(--ok)" : reach === "timeout" ? "var(--warn)" : reach === "error" ? "var(--err)" : "var(--fg-5)";
-
-  return (
-    <div className="card" style={{padding: 16, display: "flex", flexDirection: "column", gap: 8}}>
-      <div style={{display: "flex", alignItems: "center", gap: 10}}>
-        <span style={{width: 8, height: 8, borderRadius: "50%", background: dotColor}} aria-label={`endpoint ${reach}`} />
-        <div className="mono" style={{fontSize: 14, fontWeight: 500}}>{md.display_name || md.agent_id || "(unnamed)"}</div>
-      </div>
-      <div className="mono" style={{fontSize: 11, color: "var(--fg-3)"}}>{md.agent_id || "—"}</div>
-      {roles.length > 0 && (
-        <div style={{display: "flex", flexWrap: "wrap", gap: 4}}>
-          {roles.map((r, i) => <span key={i} className="chip">{r}</span>)}
-        </div>
-      )}
-      <div className="mono" style={{fontSize: 10.5, color: "var(--fg-4)"}}>
-        endpoint: {endpoint.url || "(none)"}<br />
-        registered: {hs.registered_at || "—"}
-      </div>
-      <button
-        onClick={() => setExpanded(e => !e)}
-        className="mono"
-        style={{
-          marginTop: 4,
-          padding: "4px 8px",
-          fontSize: 10,
-          background: "transparent",
-          border: "1px solid var(--line)",
-          borderRadius: 4,
-          color: "var(--fg-3)",
-          cursor: "pointer",
-          alignSelf: "flex-start",
-        }}
-      >{expanded ? "hide" : "show"} metadata</button>
-      {expanded && (
-        <pre className="mono" style={{fontSize: 10, color: "var(--fg-3)", overflow: "auto", maxHeight: 220, margin: 0, padding: 8, background: "var(--bg-2)", borderRadius: 4}}>
-          {JSON.stringify(md, null, 2)}
-        </pre>
-      )}
-    </div>
-  );
-}
-
-Object.assign(window, { BackendsView, LogsView, AgentView, AgentPeers, PeerCard });
+Object.assign(window, { BackendsView, LogsView });
